fixes
svn path=/trunk/koffice/; revision=21154

+++ kdiagramm/kdiagramm_shell.cc
@@ -1,21 +1,21 @@
 /* This file is part of the KDE project
    Copyright (C) 1998, 1999 Torben Weis <weis@kde.org>
- 
+
    This library is free software; you can redistribute it and/or
    modify it under the terms of the GNU Library General Public
    License as published by the Free Software Foundation; either
    version 2 of the License, or (at your option) any later version.
- 
+
    This library is distributed in the hope that it will be useful,
    but WITHOUT ANY WARRANTY; without even the implied warranty of
    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
    Library General Public License for more details.
- 
+
    You should have received a copy of the GNU Library General Public License
    along with this library; see the file COPYING.LIB.  If not, write to
    the Free Software Foundation, Inc., 59 Temple Place - Suite 330,
    Boston, MA 02111-1307, USA.
-*/     
+*/
 
 #include <qprinter.h>
 #include "kdiagramm_shell.h"
@@ -35,19 +35,19 @@ KDiagrammShell::KDiagrammShell()
 {
   m_pDoc = 0L;
   m_pView = 0L;
-  
+
   if ( s_lstShells == 0L )
     s_lstShells = new QList<KDiagrammShell>;
-  
+
   s_lstShells->append( this );
 }
 
 KDiagrammShell::~KDiagrammShell()
-{ 
+{
   cerr << "KDiagrammShell::~KDiagrammShell()" << endl;
-  
+
   cleanUp();
-  
+
   s_lstShells->removeRef( this );
 }
 
@@ -55,7 +55,7 @@ bool KDiagrammShell::isModified()
 {
   if ( m_pDoc )
     return (bool)m_pDoc->isModified();
-  
+
   return false;
 }
 
@@ -63,13 +63,13 @@ bool KDiagrammShell::requestClose()
 {
   int res = QMessageBox::warning( 0L, i18n("Warning"), i18n("The document has been modified\nDo you want to save it ?" ),
 				  i18n("Yes"), i18n("No"), i18n("Cancel") );
-  
+
   if ( res == 0 )
     return saveDocument( "", "" );
-  
+
   if ( res == 1 )
     return true;
-  
+
   return false;
 }
 
@@ -84,14 +84,14 @@ void KDiagrammShell::setDocument( KDiagrammDoc *_doc )
 {
   if ( m_pDoc )
     releaseDocument();
-  
+
   m_pDoc = _doc;
   m_pDoc->_ref();
   m_pView = _doc->createDiagrammView( getFrame() );
   m_pView->incRef();
   m_pView->setMode( KOffice::View::RootMode );
   m_pView->setMainWindow( interface() );
-  
+
   setRootPart( m_pView );
   interface()->setActivePart( m_pView->id() );
 
@@ -102,7 +102,7 @@ void KDiagrammShell::setDocument( KDiagrammDoc *_doc )
     m_pFileMenu->setItemEnabled( m_idMenuFile_Close, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Quit, true );
   }
-  
+
   opToolBar()->setItemEnabled( TOOLBAR_PRINT, true );
   opToolBar()->setItemEnabled( TOOLBAR_SAVE, true );
 }
@@ -116,23 +116,23 @@ bool KDiagrammShell::newDocument()
     s->newDocument();
     return true;
   }
-  
+
   m_pDoc = new KDiagrammDoc;
   if ( !m_pDoc->init() )
   {
     cerr << "ERROR: Could not initialize document" << endl;
     return false;
   }
-  
+
   m_pView = m_pDoc->createDiagrammView( getFrame() );
   m_pView->incRef();
   m_pView->setMode( KOffice::View::RootMode );
   cerr << "*1) VIEW void KOMBase::refcnt() = " << m_pView->_refcnt() << endl;
   m_pView->setMainWindow( interface() );
-  
+
   setRootPart( m_pView );
   interface()->setActivePart( m_pView->id() );
-  
+
   if( m_pFileMenu )
   {
     m_pFileMenu->setItemEnabled( m_idMenuFile_Save, true );
@@ -140,7 +140,7 @@ bool KDiagrammShell::newDocument()
     m_pFileMenu->setItemEnabled( m_idMenuFile_Close, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Quit, true );
   }
-  
+
   opToolBar()->setItemEnabled( TOOLBAR_PRINT, true );
   opToolBar()->setItemEnabled( TOOLBAR_SAVE, true );
 
@@ -162,32 +162,32 @@ bool KDiagrammShell::openDocument( const char *_url, const char *_format )
     s->show();
     return s->openDocument( _url, _format );
   }
-  
+
   cerr << "Creating new document" << endl;
-  
+
   m_pDoc = new KDiagrammDoc;
   if ( !m_pDoc->loadFromURL( _url, _format ) )
     return false;
-  
+
   m_pView = m_pDoc->createDiagrammView( getFrame() );
   m_pView->incRef();
   m_pView->setMode( KOffice::View::RootMode );
   m_pView->setMainWindow( interface() );
-  
+
   setRootPart( m_pView );
   interface()->setActivePart( m_pView->id() );
-  
+
   if ( m_pFileMenu )
-  {    
+  {
     m_pFileMenu->setItemEnabled( m_idMenuFile_SaveAs, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Save, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Close, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Quit, true );
   }
-  
+
   opToolBar()->setItemEnabled( TOOLBAR_PRINT, true );
   opToolBar()->setItemEnabled( TOOLBAR_SAVE, true );
-  
+
   return true;
 }
 
@@ -201,7 +201,7 @@ bool KDiagrammShell::saveDocument( const char *_url, const char *_format )
     url = m_pDoc->url();
     _url = url.in();
   }
-  
+
   QString file;
   if ( _url == 0L || *_url == 0 )
   {
@@ -211,10 +211,10 @@ bool KDiagrammShell::saveDocument( const char *_url, const char *_format )
       return false;
     _url = file.data();
   }
-  
+
   if ( _format == 0L || *_format == 0 )
     _format = "application/x-kdiagramm";
-  
+
   return m_pDoc->saveToURL( _url, _format );
 }
 
@@ -252,7 +252,7 @@ bool KDiagrammShell::closeAllDocuments()
 	return false;
     }
   }
-  
+
   return true;
 }
 
@@ -267,20 +267,20 @@ void KDiagrammShell::releaseDocument()
   if ( m_pDoc )
     views = m_pDoc->viewCount();
   cerr << "############## VIEWS=" << views << " #####################" << endl;
-  
+
   cerr << "-1) VIEW void KOMBase::refcnt() = " << m_pView->_refcnt() << endl;
 
-  setRootPart( (OpenParts::Id)0 );
+  setRootPart( (long unsigned int)(OpenParts::Id)0 );
 
   cerr << "-2) VIEW void KOMBase::refcnt() = " << m_pView->_refcnt() << endl;
 
   interface()->setActivePart( 0 );
 
   // cerr << "-3) VIEW void KOMBase::refcnt() = " << m_pView->_refcnt() << endl;
-  
+
   if ( m_pView )
     m_pView->decRef();
-  
+
   /* if ( m_pView )
     m_pView->cleanUp(); */
 
@@ -300,7 +300,7 @@ void KDiagrammShell::releaseDocument()
 
 void KDiagrammShell::slotFileNew()
 {
-  if ( !newDocument() )    
+  if ( !newDocument() )
     QMessageBox::critical( this, i18n("KDiagramm Error"), i18n("Could not create new document"), i18n("OK") );
 }
 
@@ -310,7 +310,7 @@ void KDiagrammShell::slotFileOpen()
 
   if ( file.isNull() )
     return;
-  
+
   if ( !openDocument( file, "" ) )
   {
     QString tmp;
@@ -322,14 +322,14 @@ void KDiagrammShell::slotFileOpen()
 void KDiagrammShell::slotFileSave()
 {
   assert( m_pDoc != 0L );
-  
+
   CORBA::String_var url = m_pDoc->url();
   if ( strlen( url.in() ) == 0 )
   {
     slotFileSaveAs();
     return;
   }
-  
+
   if ( !saveDocument( url.in(), "" ) )
   {
     QString tmp;
@@ -355,18 +355,18 @@ void KDiagrammShell::slotFileClose()
     slotFileQuit();
     return;
   }
-  
+
   if ( isModified() )
     if ( !requestClose() )
       return;
-  
+
   delete this;
 }
 
 void KDiagrammShell::slotFilePrint()
 {
   assert( m_pView );
-  
+
   (void)m_pView->printDlg();
 }
 
@@ -378,7 +378,7 @@ void KDiagrammShell::slotFileQuit()
     return;
 
   cerr << "EXIT 2" << endl;
-  
+
   delete this;
   kapp->exit();
 }
