+++ kformula/kformula_shell.cc
@@ -17,19 +17,19 @@ KFormulaShell::KFormulaShell()
 {
     m_pDoc = 0L;
     m_pView = 0L;
-  
+
     if ( s_lstShells == 0L )
 	s_lstShells = new QList<KFormulaShell>;
-  
+
     s_lstShells->append( this );
 }
 
 KFormulaShell::~KFormulaShell()
-{ 
+{
     cerr << "KFormulaShell::~KFormulaShell()" << endl;
-  
+
     cleanUp();
-  
+
     s_lstShells->removeRef( this );
 }
 
@@ -37,7 +37,7 @@ bool KFormulaShell::isModified()
 {
     if ( m_pDoc )
 	return (bool)m_pDoc->isModified();
-  
+
     return false;
 }
 
@@ -45,13 +45,13 @@ bool KFormulaShell::requestClose()
 {
     int res = QMessageBox::warning( 0L, i18n("Warning"), i18n("The document has been modified\nDo you want to save it ?" ),
 				    i18n("Yes"), i18n("No"), i18n("Cancel") );
-  
+
     if ( res == 0 )
 	return saveDocument( "", "" );
-  
+
     if ( res == 1 )
 	return true;
-  
+
     return false;
 }
 
@@ -66,14 +66,14 @@ void KFormulaShell::setDocument( KFormulaDoc *_doc )
 {
     if ( m_pDoc )
 	releaseDocument();
-  
+
     m_pDoc = _doc;
     m_pDoc->_ref();
     m_pView = _doc->createFormulaView( getFrame() );
     m_pView->incRef();
     m_pView->setMode( KOffice::View::RootMode );
     m_pView->setMainWindow( interface() );
-  
+
     setRootPart( m_pView );
     interface()->setActivePart( m_pView->id() );
 
@@ -85,7 +85,7 @@ void KFormulaShell::setDocument( KFormulaDoc *_doc )
 	    m_pFileMenu->setItemEnabled( m_idMenuFile_Print, true );
 	    m_pFileMenu->setItemEnabled( m_idMenuFile_Quit, true );
 	}
-  
+
     opToolBar()->setItemEnabled( TOOLBAR_PRINT, true );
     opToolBar()->setItemEnabled( TOOLBAR_SAVE, true );
 }
@@ -99,23 +99,23 @@ bool KFormulaShell::newDocument()
 	    s->newDocument();
 	    return true;
 	}
-  
+
     m_pDoc = new KFormulaDoc;
     if ( !m_pDoc->init() )
 	{
 	    cerr << "ERROR: Could not initialize document" << endl;
 	    return false;
 	}
-  
+
     m_pView = m_pDoc->createFormulaView( getFrame() );
     m_pView->incRef();
     m_pView->setMode( KOffice::View::RootMode );
     cerr << "*1) VIEW void KOMBase::refcnt() = " << m_pView->_refcnt() << endl;
     m_pView->setMainWindow( interface() );
-  
+
     setRootPart( m_pView );
     interface()->setActivePart( m_pView->id() );
-  
+
     if( m_pFileMenu )
 	{
 	    m_pFileMenu->setItemEnabled( m_idMenuFile_Save, true );
@@ -123,7 +123,7 @@ bool KFormulaShell::newDocument()
 	    m_pFileMenu->setItemEnabled( m_idMenuFile_Close, true );
 	    m_pFileMenu->setItemEnabled( m_idMenuFile_Quit, true );
 	}
-  
+
     opToolBar()->setItemEnabled( TOOLBAR_PRINT, true );
     opToolBar()->setItemEnabled( TOOLBAR_SAVE, true );
 
@@ -145,46 +145,46 @@ bool KFormulaShell::openDocument( const char *_url, const char *_format )
 	    s->show();
 	    return s->openDocument( _url, _format );
 	}
-  
+
     cerr << "Creating new document" << endl;
-  
+
     m_pDoc = new KFormulaDoc;
     if ( !m_pDoc->loadFromURL( _url, _format ) )
 	return false;
-  
+
     m_pView = m_pDoc->createFormulaView( getFrame() );
     m_pView->incRef();
     m_pView->setMode( KOffice::View::RootMode );
     m_pView->setMainWindow( interface() );
-  
+
     setRootPart( m_pView );
     interface()->setActivePart( m_pView->id() );
-  
+
     if ( m_pFileMenu )
-	{    
+	{
 	    m_pFileMenu->setItemEnabled( m_idMenuFile_SaveAs, true );
 	    m_pFileMenu->setItemEnabled( m_idMenuFile_Save, true );
 	    m_pFileMenu->setItemEnabled( m_idMenuFile_Close, true );
 	    m_pFileMenu->setItemEnabled( m_idMenuFile_Quit, true );
 	}
-  
+
     opToolBar()->setItemEnabled( TOOLBAR_PRINT, true );
     opToolBar()->setItemEnabled( TOOLBAR_SAVE, true );
-  
+
     return true;
 }
 
 bool KFormulaShell::saveDocument( const char *_url, const char *_format )
 {
     assert( m_pDoc != 0L );
-    
+
     CORBA::String_var url;
     if ( _url == 0L || *_url == 0 )
 	{
 	    url = m_pDoc->url();
 	    _url = url.in();
 	}
-  
+
     QString file;
     if ( _url == 0L || *_url == 0 )
 	{
@@ -195,10 +195,10 @@ bool KFormulaShell::saveDocument( const char *_url, const char *_format )
 		return false;
 	    _url = file.data();
 	}
-  
+
     if ( _format == 0L || *_format == 0 )
 	_format = "application/x-kformula";
-  
+
     return m_pDoc->saveToURL( _url, _format );
 }
 
@@ -236,7 +236,7 @@ bool KFormulaShell::closeAllDocuments()
 			return false;
 		}
 	}
-  
+
     return true;
 }
 
@@ -251,20 +251,20 @@ void KFormulaShell::releaseDocument()
     if ( m_pDoc )
 	views = m_pDoc->viewCount();
     cerr << "############## VIEWS=" << views << " #####################" << endl;
-  
+
     cerr << "-1) VIEW void KOMBase::refcnt() = " << m_pView->_refcnt() << endl;
 
-    setRootPart( (OpenParts::Id)0 );
+    setRootPart( (long unsigned int)(OpenParts::Id)0 );
 
     cerr << "-2) VIEW void KOMBase::refcnt() = " << m_pView->_refcnt() << endl;
 
     interface()->setActivePart( 0 );
 
     // cerr << "-3) VIEW void KOMBase::refcnt() = " << m_pView->_refcnt() << endl;
-  
+
     if ( m_pView )
 	m_pView->decRef();
-  
+
     /* if ( m_pView )
        m_pView->cleanUp(); */
 
@@ -284,7 +284,7 @@ void KFormulaShell::releaseDocument()
 
 void KFormulaShell::slotFileNew()
 {
-    if ( !newDocument() )    
+    if ( !newDocument() )
 	QMessageBox::critical( this, i18n("KFormula Error"), i18n("Could not create new document"), i18n("Ok") );
 }
 
@@ -294,7 +294,7 @@ void KFormulaShell::slotFileOpen()
 
     if ( file.isNull() )
 	return;
-  
+
     if ( !openDocument( file, "" ) )
 	{
 	    QString tmp;
@@ -306,14 +306,14 @@ void KFormulaShell::slotFileOpen()
 void KFormulaShell::slotFileSave()
 {
     assert( m_pDoc != 0L );
-  
+
     CORBA::String_var url = m_pDoc->url();
     if ( strlen( url.in() ) == 0 )
 	{
 	    slotFileSaveAs();
 	    return;
 	}
-  
+
     if ( !saveDocument( url.in(), "" ) )
 	{
 	    QString tmp;
@@ -339,18 +339,18 @@ void KFormulaShell::slotFileClose()
 	    slotFileQuit();
 	    return;
 	}
-  
+
     if ( isModified() )
 	if ( !requestClose() )
 	    return;
-  
+
     delete this;
 }
 
 void KFormulaShell::slotFilePrint()
 {
     assert( m_pView );
-  
+
     (void)m_pView->printDlg();
 }
 
@@ -362,7 +362,7 @@ void KFormulaShell::slotFileQuit()
 	return;
 
     cerr << "EXIT 2" << endl;
-  
+
     delete this;
     kapp->exit();
 }
