+++ kimage/kimage_shell.cc
@@ -1,21 +1,21 @@
 /* This file is part of the KDE project
    Copyright (C) 1998, 1999 Torben Weis <weis@kde.org>
- 
+
    This library is free software; you can redistribute it and/or
    modify it under the terms of the GNU Library General Public
    License as published by the Free Software Foundation; either
    version 2 of the License, or (at your option) any later version.
- 
+
    This library is distributed in the hope that it will be useful,
    but WITHOUT ANY WARRANTY; without even the implied warranty of
    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
    Library General Public License for more details.
- 
+
    You should have received a copy of the GNU Library General Public License
    along with this library; see the file COPYING.LIB.  If not, write to
    the Free Software Foundation, Inc., 59 Temple Place - Suite 330,
    Boston, MA 02111-1307, USA.
-*/     
+*/
 
 #include <qprinter.h>
 #include <qmsgbox.h>
@@ -34,26 +34,26 @@
 #include "kimage_shell.h"
 #include "kimage_doc.h"
 #include "kimage_view.h"
- 
+
 QList<KImageShell>* KImageShell::s_lstShells = 0L;
 
 KImageShell::KImageShell()
 {
   m_pDoc = 0L;
   m_pView = 0L;
-  
+
   if ( s_lstShells == 0L )
     s_lstShells = new QList<KImageShell>;
-  
+
   s_lstShells->append( this );
 }
 
 KImageShell::~KImageShell()
-{ 
+{
   debug( "KImageShell::~KImageShell()" );
-  
+
   cleanUp();
-  
+
   s_lstShells->removeRef( this );
 }
 
@@ -61,7 +61,7 @@ bool KImageShell::isModified()
 {
   if ( m_pDoc )
     return (bool)m_pDoc->isModified();
-  
+
   return false;
 }
 
@@ -69,13 +69,13 @@ bool KImageShell::requestClose()
 {
   int res = QMessageBox::warning( 0L, i18n("Warning"), i18n("The document has been modified\nDo you want to save it ?" ),
 				  i18n("Yes"), i18n("No"), i18n("Cancel") );
-  
+
   if ( res == 0 )
     return saveDocument( "", "" );
-  
+
   if ( res == 1 )
     return true;
-  
+
   return false;
 }
 
@@ -90,14 +90,14 @@ void KImageShell::setDocument( KImageDoc *_doc )
 {
   if ( m_pDoc )
     releaseDocument();
-  
+
   m_pDoc = _doc;
   m_pDoc->_ref();
   m_pView = _doc->createImageView( getFrame() );
   m_pView->incRef();
   m_pView->setMode( KOffice::View::RootMode );
   m_pView->setMainWindow( interface() );
-  
+
   setRootPart( m_pView );
   interface()->setActivePart( m_pView->id() );
 
@@ -108,7 +108,7 @@ void KImageShell::setDocument( KImageDoc *_doc )
     m_pFileMenu->setItemEnabled( m_idMenuFile_Close, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Print, true );
   }
-  
+
   opToolBar()->setItemEnabled( TOOLBAR_PRINT, true );
   opToolBar()->setItemEnabled( TOOLBAR_SAVE, true );
 }
@@ -122,7 +122,7 @@ bool KImageShell::newDocument()
     s->newDocument();
     return true;
   }
-  
+
   m_pDoc = new KImageDoc;
   if ( !m_pDoc->init() )
   {
@@ -130,16 +130,16 @@ bool KImageShell::newDocument()
     debug( "ERROR: Could not initialize document" );
     return false;
   }
-  
+
   m_pView = m_pDoc->createImageView( getFrame() );
   m_pView->incRef();
   m_pView->setMode( KOffice::View::RootMode );
   debug( "*1) VIEW void KOMBase::refcnt() = %li", m_pView->_refcnt() );
   m_pView->setMainWindow( interface() );
-  
+
   setRootPart( m_pView );
   interface()->setActivePart( m_pView->id() );
-  
+
   if( m_pFileMenu )
   {
     m_pFileMenu->setItemEnabled( m_idMenuFile_Save, true );
@@ -147,7 +147,7 @@ bool KImageShell::newDocument()
     m_pFileMenu->setItemEnabled( m_idMenuFile_Close, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Print, true );
   }
-  
+
   opToolBar()->setItemEnabled( TOOLBAR_PRINT, true );
   opToolBar()->setItemEnabled( TOOLBAR_SAVE, true );
 
@@ -169,32 +169,32 @@ bool KImageShell::openDocument( const char *_url, const char *_format )
     s->show();
     return s->openDocument( _url, _format );
   }
-  
+
   debug( "Creating new document" );
-  
+
   m_pDoc = new KImageDoc;
   if ( !m_pDoc->loadFromURL( _url, _format ) )
     return false;
-  
+
   m_pView = m_pDoc->createImageView( getFrame() );
   m_pView->incRef();
   m_pView->setMode( KOffice::View::RootMode );
   m_pView->setMainWindow( interface() );
-  
+
   setRootPart( m_pView );
   interface()->setActivePart( m_pView->id() );
-  
+
   if ( m_pFileMenu )
-  {    
+  {
     m_pFileMenu->setItemEnabled( m_idMenuFile_SaveAs, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Save, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Close, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Print, true );
   }
-  
+
   opToolBar()->setItemEnabled( TOOLBAR_PRINT, true );
   opToolBar()->setItemEnabled( TOOLBAR_SAVE, true );
-  
+
   return true;
 }
 
@@ -208,7 +208,7 @@ bool KImageShell::saveDocument( const char *_url, const char *_format )
     url = m_pDoc->url();
     _url = url.in();
   }
-  
+
   QString file;
   if ( _url == 0L || *_url == 0 )
   {
@@ -218,10 +218,10 @@ bool KImageShell::saveDocument( const char *_url, const char *_format )
       return false;
     _url = file.data();
   }
-  
+
   if ( _format == 0L || *_format == 0 )
     _format = "application/x-kimage";
-  
+
   return m_pDoc->saveToURL( _url, _format );
 }
 
@@ -259,7 +259,7 @@ bool KImageShell::closeAllDocuments()
 	return false;
     }
   }
-  
+
   return true;
 }
 
@@ -274,13 +274,13 @@ void KImageShell::releaseDocument()
   if ( m_pDoc )
     views = m_pDoc->viewCount();
 
-  setRootPart( (OpenParts::Id)0 );
+  setRootPart( (long unsigned int)(OpenParts::Id)0 );
 
   interface()->setActivePart( 0 );
 
   if ( m_pView )
     m_pView->decRef();
-  
+
   /* if ( m_pView )
     m_pView->cleanUp(); */
 
@@ -300,14 +300,14 @@ void KImageShell::releaseDocument()
     m_pFileMenu->setItemEnabled( m_idMenuFile_Close, false );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Print, false );
   }
- 
+
   opToolBar()->setItemEnabled( TOOLBAR_PRINT, false );
   opToolBar()->setItemEnabled( TOOLBAR_SAVE, false );
 }
 
 void KImageShell::slotFileNew()
 {
-  if ( !newDocument() )    
+  if ( !newDocument() )
     QMessageBox::critical( this, i18n("KImage Error"), i18n("Could not create new document"), i18n("OK") );
 }
 
@@ -317,7 +317,7 @@ void KImageShell::slotFileOpen()
 
   if ( file.isNull() )
     return;
-  
+
   if ( !openDocument( file, "" ) )
   {
     QString tmp;
@@ -329,14 +329,14 @@ void KImageShell::slotFileOpen()
 void KImageShell::slotFileSave()
 {
   assert( m_pDoc != 0L );
-  
+
   CORBA::String_var url = m_pDoc->url();
   if ( strlen( url.in() ) == 0 )
   {
     slotFileSaveAs();
     return;
   }
-  
+
   if ( !saveDocument( url.in(), "" ) )
   {
     QString tmp;
@@ -360,14 +360,14 @@ void KImageShell::slotFileClose()
   if ( isModified() )
     if ( !requestClose() )
       return;
-  
+
   releaseDocument();
 }
 
 void KImageShell::slotFilePrint()
 {
   assert( m_pView );
-  
+
   (void)m_pView->printDlg();
 }
 
@@ -379,7 +379,7 @@ void KImageShell::slotFileQuit()
     return;
 
   debug( "EXIT 2" );
-  
+
   delete this;
   kapp->exit();
 }
