+++ kspread/kspread_shell.cc
@@ -1,21 +1,21 @@
 /* This file is part of the KDE project
    Copyright (C) 1998, 1999 Torben Weis <weis@kde.org>
- 
+
    This library is free software; you can redistribute it and/or
    modify it under the terms of the GNU Library General Public
    License as published by the Free Software Foundation; either
    version 2 of the License, or (at your option) any later version.
- 
+
    This library is distributed in the hope that it will be useful,
    but WITHOUT ANY WARRANTY; without even the implied warranty of
    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
    Library General Public License for more details.
- 
+
    You should have received a copy of the GNU Library General Public License
    along with this library; see the file COPYING.LIB.  If not, write to
    the Free Software Foundation, Inc., 59 Temple Place - Suite 330,
    Boston, MA 02111-1307, USA.
-*/     
+*/
 
 #include <qprinter.h>
 #include "kspread_shell.h"
@@ -37,19 +37,19 @@ KSpreadShell::KSpreadShell()
 {
   m_pDoc = 0L;
   m_pView = 0L;
-  
+
   if ( s_lstShells == 0L )
     s_lstShells = new QList<KSpreadShell>;
-  
+
   s_lstShells->append( this );
 }
 
 KSpreadShell::~KSpreadShell()
-{ 
+{
   cerr << "KSpreadShell::~KSpreadShell()" << endl;
-  
+
   cleanUp();
-  
+
   s_lstShells->removeRef( this );
 }
 
@@ -57,7 +57,7 @@ bool KSpreadShell::isModified()
 {
   if ( m_pDoc )
     return (bool)m_pDoc->isModified();
-  
+
   return false;
 }
 
@@ -65,13 +65,13 @@ bool KSpreadShell::requestClose()
 {
   int res = QMessageBox::warning( 0L, i18n("Warning"), i18n("The document has been modified\nDo you want to save it ?" ),
 				  i18n("Yes"), i18n("No"), i18n("Cancel") );
-  
+
   if ( res == 0 )
     return saveDocument( "", "" );
-  
+
   if ( res == 1 )
     return true;
-  
+
   return false;
 }
 
@@ -86,14 +86,14 @@ void KSpreadShell::setDocument( KSpreadDoc *_doc )
 {
   if ( m_pDoc )
     releaseDocument();
-  
+
   m_pDoc = _doc;
   m_pDoc->_ref();
   m_pView = _doc->createSpreadView( getFrame() );
   m_pView->incRef();
   m_pView->setMode( KOffice::View::RootMode );
   m_pView->setMainWindow( interface() );
-  
+
   setRootPart( m_pView );
   interface()->setActivePart( m_pView->id() );
 
@@ -104,7 +104,7 @@ void KSpreadShell::setDocument( KSpreadDoc *_doc )
     m_pFileMenu->setItemEnabled( m_idMenuFile_Close, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Quit, true );
   }
-  
+
   opToolBar()->setItemEnabled( TOOLBAR_PRINT, true );
   opToolBar()->setItemEnabled( TOOLBAR_SAVE, true );
   opToolBar()->setFullWidth(false);
@@ -119,23 +119,23 @@ bool KSpreadShell::newDocument()
     s->newDocument();
     return true;
   }
-  
+
   m_pDoc = new KSpreadDoc;
   if ( !m_pDoc->init() )
   {
     cerr << "ERROR: Could not initialize document" << endl;
     return false;
   }
-  
+
   m_pView = m_pDoc->createSpreadView( getFrame() );
   m_pView->incRef();
   m_pView->setMode( KOffice::View::RootMode );
   cerr << "*1) VIEW void KOMBase::refcnt() = " << m_pView->_refcnt() << endl;
   m_pView->setMainWindow( interface() );
-  
+
   setRootPart( m_pView );
   interface()->setActivePart( m_pView->id() );
-  
+
   if( m_pFileMenu )
   {
     m_pFileMenu->setItemEnabled( m_idMenuFile_Save, true );
@@ -143,7 +143,7 @@ bool KSpreadShell::newDocument()
     m_pFileMenu->setItemEnabled( m_idMenuFile_Close, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Quit, true );
   }
-  
+
   opToolBar()->setItemEnabled( TOOLBAR_PRINT, true );
   opToolBar()->setItemEnabled( TOOLBAR_SAVE, true );
   opToolBar()->setFullWidth(false);
@@ -166,9 +166,9 @@ bool KSpreadShell::openDocument( const char *_url, const char *_format )
     s->show();
     return s->openDocument( _url, _format );
   }
-  
+
   cerr << "Creating new document" << endl;
-  
+
   m_pDoc = new KSpreadDoc;
   if ( !m_pDoc->loadFromURL( _url, _format ) )
   {
@@ -176,23 +176,23 @@ bool KSpreadShell::openDocument( const char *_url, const char *_format )
     m_pDoc = 0L;
     return false;
   }
-  
+
   m_pView = m_pDoc->createSpreadView( getFrame() );
   m_pView->incRef();
   m_pView->setMode( KOffice::View::RootMode );
   m_pView->setMainWindow( interface() );
-  
+
   setRootPart( m_pView );
   interface()->setActivePart( m_pView->id() );
-  
+
   if ( m_pFileMenu )
-  {    
+  {
     m_pFileMenu->setItemEnabled( m_idMenuFile_SaveAs, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Save, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Close, true );
     m_pFileMenu->setItemEnabled( m_idMenuFile_Quit, true );
   }
-  
+
   opToolBar()->setItemEnabled( TOOLBAR_PRINT, true );
   opToolBar()->setItemEnabled( TOOLBAR_SAVE, true );
   opToolBar()->setFullWidth(false);
@@ -210,7 +210,7 @@ bool KSpreadShell::saveDocument( const char *_url, const char *_format )
     url = m_pDoc->url();
     _url = url.in();
   }
-  
+
   QString file;
   if ( _url == 0L || *_url == 0 )
   {
@@ -220,10 +220,10 @@ bool KSpreadShell::saveDocument( const char *_url, const char *_format )
       return false;
     _url = file.data();
   }
-  
+
   if ( _format == 0L || *_format == 0 )
     _format = "application/x-kspread";
-  
+
   return m_pDoc->saveToURL( _url, _format );
 }
 
@@ -261,7 +261,7 @@ bool KSpreadShell::closeAllDocuments()
 	return false;
     }
   }
-  
+
   return true;
 }
 
@@ -276,20 +276,20 @@ void KSpreadShell::releaseDocument()
   if ( m_pDoc )
     views = m_pDoc->viewCount();
   cerr << "############## VIEWS=" << views << " #####################" << endl;
-  
+
   cerr << "-1) VIEW void KOMBase::refcnt() = " << m_pView->_refcnt() << endl;
 
-  setRootPart( (OpenParts::Id)0 );
+  setRootPart( (long unsigned int)(OpenParts::Id)0 );
 
   cerr << "-2) VIEW void KOMBase::refcnt() = " << m_pView->_refcnt() << endl;
 
   interface()->setActivePart( 0 );
 
   // cerr << "-3) VIEW void KOMBase::refcnt() = " << m_pView->_refcnt() << endl;
-  
+
   if ( m_pView )
     m_pView->decRef();
-  
+
   /* if ( m_pView )
     m_pView->cleanUp(); */
 
@@ -309,7 +309,7 @@ void KSpreadShell::releaseDocument()
 
 void KSpreadShell::slotFileNew()
 {
-  if ( !newDocument() )    
+  if ( !newDocument() )
     QMessageBox::critical( this, i18n("KSpread Error"), i18n("Could not create new document"), i18n("Ok") );
 }
 
@@ -319,7 +319,7 @@ void KSpreadShell::slotFileOpen()
 							      "application/x-kspread",
 							      "*.ksp", "KSpread",
 							      true );
-      
+
   QString file = KFileDialog::getOpenFileName( getenv( "HOME" ), filter );
   if ( file.isNull() )
     return;
@@ -327,7 +327,7 @@ void KSpreadShell::slotFileOpen()
   file = KoFilterManager::self()->import( file, "application/x-kspread" );
   if ( file.isNull() )
     return;
-  
+
   if ( !openDocument( file, "" ) )
   {
     QString tmp( i18n( "Could not open\n%s" ).arg( file.data() ) );
@@ -338,14 +338,14 @@ void KSpreadShell::slotFileOpen()
 void KSpreadShell::slotFileSave()
 {
   assert( m_pDoc != 0L );
-  
+
   CORBA::String_var url = m_pDoc->url();
   if ( strlen( url.in() ) == 0 )
   {
     slotFileSaveAs();
     return;
   }
-  
+
   if ( !saveDocument( url.in(), "" ) )
   {
     QString tmp;
@@ -371,18 +371,18 @@ void KSpreadShell::slotFileClose()
     slotFileQuit();
     return;
   }
-  
+
   if ( isModified() )
     if ( !requestClose() )
       return;
-  
+
   delete this;
 }
 
 void KSpreadShell::slotFilePrint()
 {
   assert( m_pView );
-  
+
   (void)m_pView->printDlg();
 }
 
@@ -394,7 +394,7 @@ void KSpreadShell::slotFileQuit()
     return;
 
   cerr << "EXIT 2" << endl;
-  
+
   delete this;
   kapp->exit();
 }
